Grouped all modules to modules directory.
