+++ modules/blog/controller/Blog.js
@@ -1,6 +1,6 @@
 exports.index = function (req, res) {
 	var responseObj = {
-		message: "Hello, from Blog index!!!"
+		message: "Hello, from Blog index :)!!!"
 	};
 	return responseObj;
 };
